Itinerary Creator: Fix itinerary order bug, write unit tests.
While the itinerary is being constructed, later stops appear towards
the beginning of the list and earlier stops towards the end. This is
done in order to take advantage of the fact that a prepend operation
on an immutable list takes O(1) time.  When the itinerary is complete,
it is reversed before returning. All functions that use the itinerary
while it is still being created should take this into account.